frontends: python: automatically identify package (#586)
* frontends: python: automatically identify package

This is an effort to remote usage of the PYFUZZPACKAGE env variable
which is currently used to point PYCG on the source to include in
the analysis.

This is done by having an individual run of each individual fuzzer's
AST, determine the imports, and determining where this import is
loaded from. Then, use this information to point to the source on the
system.

This is wip. Some missing pieces are here that needs to be updated in
the oss-fuzz diff. This will come when it's more stable.

At the moment, this work with the "glom" project, in that we dont have
to add the PYFUZZPACKAGE anymore there. However, a bit more logic is
needed for it to work on more (all, fingers crossed) packages.

Signed-off-by: David Korczynski <david@adalogics.com>

* update automatic package identification

Signed-off-by: David Korczynski <david@adalogics.com>

* add oss-fuzz patches

Signed-off-by: David Korczynski <david@adalogics.com>

* clenaup oss-fuzz patches

Signed-off-by: David Korczynski <david@adalogics.com>

* nits

Signed-off-by: David Korczynski <david@adalogics.com>

* cleanup style

Signed-off-by: David Korczynski <david@adalogics.com>

* nits

Signed-off-by: David Korczynski <david@adalogics.com>

* fix styling

Signed-off-by: David Korczynski <david@adalogics.com>

* fix styling

Signed-off-by: David Korczynski <david@adalogics.com>

* update PyCG dependency

Signed-off-by: David Korczynski <david@adalogics.com>

Signed-off-by: David Korczynski <david@adalogics.com>