python-logic: improve import handling in frontend and refine link handling (#1006)
* python-logic: improve import handling in frontend and refine link
handling

A number of projects are failing to have any code analysed from the
giving libraries they represent. This includes e.g. netaddr-py
https://storage.googleapis.com/oss-fuzz-introspector/netaddr-py/inspector-report/20230423/fuzz_report.html

This PR addresses the above issue due to a bug in limiting the python
modules when running the fuzz introspector frontend logic. The problem
is, in order to identify what code to anlyse we rely on some AST parsing
for capturing relevant import modules, and there was an issue in this
code which limited the modules we'd include.

Also improves the way we create links in HTML reports for python
projects.

Signed-off-by: David Korczynski <david@adalogics.com>

* nit

Signed-off-by: David Korczynski <david@adalogics.com>

* nit

Signed-off-by: David Korczynski <david@adalogics.com>

* switch to debug logging

Signed-off-by: David Korczynski <david@adalogics.com>

---------

Signed-off-by: David Korczynski <david@adalogics.com>